requirements.txt
# Free-Rider Collective Intelligence Dependencies
# Core Memo DSL and probabilistic programming
memo-lang>=1.2.1

# Scientific computing and analysis
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter notebook support
jupyter>=1.0.0
jupyterlab>=3.0.0
jupytext>=1.13.0

# Statistical analysis
statsmodels>=0.12.0

# Type hints and development
typing-extensions>=4.0.0

# Optional: Enhanced notebook experience
ipywidgets>=7.6.0
plotly>=5.0.0
